Moved PlatformManager and related modules to 'platforms' package
