+++ dpl/connections/__init__.py
@@ -1,5 +1,3 @@
 from .connection import Connection
-from .connection_factory import ConnectionFactory
-from .connection_registry import ConnectionRegistry
 
-__all__ = ['Connection', 'ConnectionFactory', 'ConnectionRegistry']
+__all__ = ['Connection']
