+++ dpl/platforms/__init__.py
@@ -1,0 +1,9 @@
+from .connection_factory import ConnectionFactory
+from .connection_registry import ConnectionRegistry
+from .thing_factory import ThingFactory
+from .thing_registry import ThingRegistry
+from .platform_manager import PlatformManager
+
+__all__ = ['ConnectionFactory', 'ConnectionRegistry', 'ThingFactory', 'ThingRegistry',
+           'PlatformManager']
+
