+++ dpl/platforms/connection_registry.py
