+++ dpl/platforms/dummy/dummy_connection.py
@@ -4,7 +4,8 @@ import io
 
 # Include 3rd-party modules
 # Include DPL modules
-from dpl.connections import Connection, ConnectionFactory, ConnectionRegistry
+from dpl.connections import Connection
+from dpl.platforms import ConnectionFactory, ConnectionRegistry
 
 
 class DummyConnection(Connection):
