+++ dpl/platforms/dummy/dummy_player.py
@@ -2,7 +2,8 @@
 # Include 3rd-party modules
 
 # Include DPL modules
-from dpl.things import Player, ThingFactory, ThingRegistry
+from dpl.things import Player
+from dpl.platforms import ThingFactory, ThingRegistry
 from . import DummyConnection
