+++ dpl/platforms/thing_registry.py
