+++ dpl/things/__init__.py
@@ -1,9 +1,7 @@
 from .thing import Thing
-from .thing_factory import ThingFactory
-from .thing_registry import ThingRegistry
 from .actuator import Actuator
 from .switch import Switch
 from .slider import Slider
 from .player import Player
 
-__all__ = ["Thing", "ThingFactory", "ThingRegistry", "Actuator", "Switch", "Slider", "Player"]
+__all__ = ["Thing", "Actuator", "Switch", "Slider", "Player"]
